arc-setup: Prevent mounts of symlinks for sdcard
arc-sdcard is an upstart job which mounts the arc ESDFS to
/run/arc/sdcard/{default,read,write}/emulated. With uid=0
code execution, it is possible to create a symlink from
/etc/init to anyone of the previous 3 locations. When the
upstart job is triggered after this, it is possible to
mount /opt/google/containers/android/rootfs/android-data/data/media
on /etc/init. This can further lead to installing another
malicious upstart service and triggering it via initctl.

This CL stops the above behavior by checking that the
Realpath of the destination always falls under the
/run/arc/sdcard directory, rather than /etc/init or something
else as in the example above.

BUG=chromium:1168553
TEST=manual

Change-Id: If299ce055c4852547d3e8c450a26934840628ef1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3224805
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Josh Horwich <jhorwich@chromium.org>
Tested-by: Vaibhav Raheja <vraheja@chromium.org>
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>