arc: Set up ARCVM's /data/media/0/Download with strict permission
Currently, users can manipulate ARCVM's Download directory before it is
covered by Chrome OS Downloads when
  1) Their device are newly opted-in, and
  2) ARCVM's sdcard volume (/data/media) is mounted, but
  3) Chrome OS Downloads is not yet bind-mounted on ARCVM's Download
     (/data/media/0/Download) on the guest side.
On such an occasion, /data/media/0/Download is owned by Android's
MediaProvider and files can be created in the directory.
This is because MediaProvider tries to create the Download directory
if it does not exist after the sdcard volume is mounted.

This CL prevents this scenario by setting up /data/media/0/Download in
arc-create-data with a strict permission so that Android's MediaProvider
won't create it. With this change, there won't be a window period in
which users can create files in the directory before it's covered by
Chrome OS Downloads.
Note that it's hard to do the bind-mount on /data/media/0/Download
earlier to avoid this race because 1) we need to set up the Download
directory on the host side using the MediaProvider UID before the
bind-mount, and 2) it's hard to guarantee that the directory is set up
before the sdcard volume is mounted on the guest side.
This CL provides a not strict but simpler solution for the issue.

BUG=b:203171528
TEST=emerge & cros deploy {arc-setup, selinux-policy}
TEST=Confirm that Downloads integration properly works after 1) upgrade,
TEST=2) opt-in, and 3) ordinary ARCVM boot.
TEST=Disable Downloads > Download mount on the Android side, and confirm
TEST=that files cannot be downloaded by Google Photos after 1) upgrade,
TEST=2) opt-in, and 3) ordinary ARCVM boot.

Change-Id: I91ec12cc0c5c74d9378f3352df11128991e8c75b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3225765
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Qijiang Fan <fqj@google.com>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Risan <risan@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>