bootid-logger: Allow very large input dates
To make this simpler to fuzz, lets allow large dates to be input as the
current time. This makes the code more resilient to the system clock
being way off so that it doesn't crash unless the format of the string
is less than what is expected. If the large format is written into the
log we'll ignore it anyway when reading logs from the file because the
dashes aren't in the right place.

BUG=None
TEST=emerge-amd64-generic bootid-logger

Change-Id: I10abd3b0cbd87cae69833fb05971d2e2363702ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3230483
Reviewed-by: Yoshiki Iguchi <yoshiki@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>