biod: Use RecordId as key in records_
Until now, records_ was a vector which contained metadata of records
that was actually loaded into FPMCU. The index was just number of
template loaded into FPMCU, this was sufficient when records_ was in
CrosFpBiometricsManager class, but now we are going to move records
related code into separate class and in that case it seems more natural
to access records metadata using RecordId.

This patch prepares CrosFpBiometricsManager class to split it into two
classes.

BUG=b:184843581
TEST=FEATURES=test emerge-hatch biod

Change-Id: Ia999d5c9d2e3ec62b1509e744bc68b3d84e27e0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3247183
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Patryk Duda <patrykd@google.com>