biod: Reorganize code which checks Positive Match Secret
Until now, DoMatchEvent used CalculateMatches to get list of records
that matched. CalculateMatches used ValidationValueIsCorrect to get
Positive Match Secret from FPMCU and check if it matches value stored on
disk.

Actually, CalculateMatches is checking if provided index is valid and
calls ValidationValueIsCorrect, to check if Positive Match Secret is
correct. Also for one matched template there is only one record that
will be returned, so it can return bool.

In this patch, validating index was moved to DoMatchEvent and
CheckSecureSecret is responsible for providing information if
Positive Match Secret is correct.

BUG=b:184843581
TEST=FEATURES=test emerge-hatch biod

Change-Id: Ied22b801e615b2d970835e83461bfb6005a7312a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3247182
Tested-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Patryk Duda <patrykd@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>