cryptohome: Modularize GetOrCreateSalt to support multiple salts
There are a few structural changes made as part of this CL:
1) GetorCreateSalt's function parameters are redecuded to FilePath and
SecureBlob*. The two removed parameters are the same everywhere so the
function now uses constants.
2) GetOrCreateSalt is marked private in favour of GetPublicMountSalt
and GetSystemSalt.
3) AssignSalt in UserDataAuth unit tests was using the same salt
size, so the function parameter is reduced to a constant.
4) kImageDir is a constant in crypto_unittest that no longer is being
used, so that is removed.

BUG=b:195660730
TEST=FEATURES=test emerge-zork cryptohome

Change-Id: Iffc9cf546bfb57ab692762a41c380d52ee88d68d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3235806
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>