power: Use elogtool for mosys_eventlog setuid helper
"mosys eventlog" is deprecated and replaced with "elogtool".  The
command-line API of "elogtool add" is compatible with "mosys eventlog
add".  This changes the mosys_eventlog action in the setuid helper to
use elogtool instead of mosys.

Note: Follow up CLs do further renaming so that the name "mosys" is
used less in the configs/APIs.  This CL just does the bare minimum to
switch over to elogtool.

BUG=b:203777035
TEST=CQ

Change-Id: If5bb1905840e40fa585273349c1478e196e658df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3237568
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>