dns-proxy: Limit process restarts
The controller is responsible for monitoring proxy child processes and
restarting them if they crash. However, it's possible that a fundamental
fatal condition may prevent the process from ever successfully starting.
This patch adds a retry guard to avoid endlessly attempting to retart in
this case:

The first time a process is restarted, the controller will remember the
time stamp and initialize a counter. On any subsequent retry, if the
attempt is made within a time span (20s) from the last attempt, the
count will be incremented. If this count exceeds a limit (10) then there
will be no retry attempt and that proxy will be marked as "blocked" and
not restarted until the daemon is (e.g. reboot).

BUG=b:201244524
TEST=verified DNS works as usual
TEST=added a condition in minijail to fail child launch, observed retry
attempts and eventual stopping thereof
TEST=same as above, ran ARC, logged out, logged back in and observed
that ARC proxies not restarted (expected)

Change-Id: I94cf7b7ec31b99a3886b9d0c7d8322c5afd32225
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3239284
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Garrick Evans <garrick@chromium.org>