ml: use approximate equality for text suggester tests
We are currently in the process of enabling the text suggestion model
for all boards (excluding 2gb boards). This includes a number of arm, x86_64
and arm64 devices.

The current integration tests written for the text suggester model have
been written for and run on x86_64 devices. When running the same
integration tests on arm boards the scoring emitted for candidates is
approximately the same but not EXACTLY equal.

In light of this upcoming change, this CL updates the integration tests
to use an EXPECT_NEAR assertion for candidate scoring with a allowable
delta of 0.0015 which covers both arm and x86_64 boards.

BUG=chromium:1146266
TEST=cros_run_unit_tests --board=${BOARD} --packages chromeos-base/ml
TEST=The above line has been run on the following boards;
TEST=nocturne -> x86_64
TEST=strongbad -> arm
TEST=trogdor64 -> arm64

Change-Id: Ic04568493e66c2f1fb9403bc21fa71779c5bebec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3239291
Tested-by: Curtis McMullan <curtismcmullan@chromium.org>
Commit-Queue: Curtis McMullan <curtismcmullan@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>