cryptohome: make LoopDeviceManager a part of Platform.
This is essentially an object which we need a singular instance of but
for now we pass it to the backing device directly. This complicates the
e2e test which mount_unittest currently is. This also may be a better
approach, considering it is a part of platform interaction, and Platform
may soon become a true singleton.

BUG=b:172344610,b:172622448
TEST=FEATURES="test" USE="test" emerge-hatch cryptohome

Change-Id: I5cb7deae423796420341a11adaa304b28b4319f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3249700
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>