shill: vpn: Report cipher suites in L2TPIPsecDriver
This patch adds the logic to report cipher suites of IKE and ESP in a
connected IPsec connection in L2TPIPsecDriver. Different from what we're
doing in IPsecConnection, this report process is async and will not
block the connect procedure. It is implemented by executing `stroke
statusall` and parse its output.

Also make ParseCipherSuite() a static member function in
IPsecConnection to make it accessible in L2TPIPsecDriver. The outputs of
stroke and swanctl for a cipher suites are the same, except for stroke
uses "_" to connect the algorithm name and key length, while swanctl
uses "-" for that. We have already unified them in the parsing
functions.

BUG=b:202795142
TEST=unit_tests
TEST=manually verified that the added metrics shown on the page of
  chrome://histograms

Change-Id: I4fb9f3ea40d46642fa95dff5487e8831357a34e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3250078
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>