cryptohome: hardcode the uids and gids
Dynamically querying those is an overkill - they are very unlikely to
change, since the change is going to be a massive migration, and if
needed querying can be easily re-introduced at any point. And doing the
hardcoding allows us to remove a lot of unnecessary code.

With that we also replace kMountOwnerUID/GID with kRootUid/Gid, for they
are the same.

BUG=b:172344610
TEST=FEATURES="test" USE="test" emerge-hatch cryptohome

Change-Id: Id37949186e3f50461091155e181abd0eb5091491
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3252996
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>