chaps: Cache result of IsTPMAvailable in TPMUtility
Currently TPMUtilityImpl doesn't cache the result of IsTPMAvailable, and
TPM2UtilityImpl partially caches the result (it caches the result of
Init, and uses it in IsTPMAvailable). This means that the code logic
doesn't guarantee once IsTPMAvailable returns true, it will never return
false in subsequent calls. This behavior might cause threading error in
SlotManagerImpl (joins a thread twice), since it decides threading
behavior using the return value of IsTPMAvailable. This bug is found by
a locally-run fuzzer. If the result is cached, once IsTPMAvailable
returns true it will always return true, which is more reasonable and
makes the call site code safer. Also, this saves a number of tripes over
dbus, since it eliminates multiple calls to GetTpmNonSensitiveStatus.

Thus, cache the result of IsTPMAvailable in both TPMUtilityImpl and
TPM2UtilityImpl.

BUG=b:172210084
TEST=CQ

Change-Id: I7272c7fb80d4391072d0b04937a6d53a644f0656
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3253006
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>