shill: vpn: Report connection end reason in L2TP/IPsec drivers
This patch adds code in the two L2TP/IPsec driver to report the reason
why a VPN connection is ended. This metric is reported once and only
once for each L2TP/IPsec VPN connection, no matter it fails during
connecting or is closed by the user.

The added metrics reuse the NetworkServiceError enum and thus the
corresponding converting code is extracted and exposed as a static
method in the Metrics class.

BUG=b:204261554
TEST=unit_tests
TEST=manually tested with both of the drivers that the corresponding
  metrics were added on chrome://histograms

Change-Id: I27a047012e9653c2e58f2521605b869a625183fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3256602
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>