vm_tools: ManaTEE memory service integration
A memory service in the manatee hypervisor is responsible for ensuring
that the hypervisor does not run out of memory. This change integrates
concierge with that memory service. This integration involves calls to
the hypervisor when launching and shutting down VMs, as well as when
rebalancing memory with sibling VMs. Memory policy decisions continue to
be made by concierge.

BUG=b:213962590
TEST=manual testing with crrev.com/c/3394436

Change-Id: I0fd9871588aa076688d62497a6bd8dcf18bfadd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3394431
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>