libhwsec-foundation: Add Synchronized class
Add Synchronized class to provide thread-safe access to thread-unsafe
objects. Most of the time, synchronization mechanisms should be
considered when designing a new class, and those classes that aren't
thread-safe shouldn't be mutably accessed in different threads. However,
sometimes we use built-in or third-party classes that do not provide
synchronization mechanisms, but we have to use it in multiple threads
(for example, FuzzedDataProvider). Synchronized is a convenient wrapper
that can be used in those cases.

Synchronized is a synchronization wrapper that owns a class object, and
the Lock() method can be used to obtain a SynchronizedHandler, which has
exclusive access of the underlying object.

BUG=None
TEST=the new unittest synchronized_test.cc

Change-Id: I9d776f973d8d6dee746cda0b4b53e4bd394e2b95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3257582
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>