chaps: Early return when token is not accessible in SlotManager::UnloadToken
Currently, in SlotManager::UnloadToken, if IsTokenAccessible returns
false, it produces a warning and continues the function trying to unload
a token. But afterwards it calls CloseAllSessions, which asserts that
IsTokenAccessible is true, which might cause the thread to die.
Therefore, early return when IsTokenAccessible returns false in
UnloadToken. This bug is found by a locally-run fuzzer.

BUG=b:172210084
TEST=CQ

Change-Id: If7f25b184b89f824e88982cf481d81e7e357c189
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3257424
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>