debugd: Add status check for packet capture process
Send the write end of a pipe to the capture_packets helper process from
the PacketCaptureTool to check if the helper process started packet
captures successfully. PacketCaptureTool::Start() will only return true
if the helper process wrote "1" on the pipe to indicate the successful
start. This will prevent the callers of the DBus method from assuming
the process already started when it didn't.
Send SendPacketCaptureStopSignal() function as a callback to the
PacketCaptureTool::Start() function. PacketCaptureTool will run the
callback when the pipe is closed when all packet capture processes are
stopped. Before this change, the signal was only emitted when packet
capture process was ended by the user. It wasn't emitted when process
terminated due to an error.
Support correct behaviour in case of multiple packet capture processes
running at the same time.

BUG=b:198759008
TEST=Run "tast -verbose run <DUT ip address> debugd.PacketCapture"

Change-Id: I2a2e8e185c9c50e48cf9a3e9ded0c006b069eb21
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3259543
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Irem Uguz <iremuguz@google.com>
Commit-Queue: Irem Uguz <iremuguz@google.com>