init: Start to move chromeos_startup to C++
This initial CL creates renames the existing bash script to
chromeos_startup.sh which is called in a brillo ProcessImpl
by the new C++ chromeos_startup executable. This CL also includes
the parsing of the USE flags and initializing the variables that
are defined in the constructor for the class. They are not all used yet
in the CL, but are required for the CLs that follow.

Design Doc for additional information:
https://docs.google.com/document/d/1VmXlvghuPwEdoxhb3pzYfYbHvOUjFmEvP0SnH3_C3tg/preview

BUG=b:172210918
TEST=build_image --board=hatch, unit tests

Cq-Depend: chromium:2999416, chromium:2288174
Change-Id: I4ce09c7bc3d8ff3374437824c1b4923e6aa5e9b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2999415
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alexis Savery <asavery@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>