biod: Refactor ReadValidationValueFromRecord in BiodStorage
Until now ReadValidationValueFromRecord was responsible for
checking if record format version is correct and deciding if records
with older versions are valid. Decision if record is valid (all fields
required by version are present) should be made during processing
record, not when reading validation value.

If a record uses old format (but it is valid record), then BiodStorage
should accept this record and let upper layers decide if the record
needs to be migrated or deleted.

BUG=b:203599040 b:184843581
TEST=FEATURES=test emerge-hatch biod

Change-Id: Ic23274edeab5a73deb42dc314e9f54d28eefc1f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3259546
Tested-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Patryk Duda <patrykd@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>