missive: Fix Scheduler, refactor its tests
In the Schedule there was a potential deep/infinite recursion:
StartJobs triggers OnPopJob callbacks which in turn call StartJobs.
This may be responsible for high CPU utilization in Brye and
is eliminated by this CL.
To fix that, asynchronous shared queue was replaced by synchronous
std::queue, so that it is now possible to not acquire semaphore if
there are no more jobs in the queue. Also, when a job completes,
instead of releasing the semaphore and calling StartJobs, we reuse
it, if there are any more jobs in the queue.

And there was a bug in scheduler - all jobs were started on the same
sequenced_task_runner, so if they did not post actual work elsewhere,
it was serialized unnecessarily.

Also, Scheduler tests are refactored, among other things by
using util/test_callbacks and adding simulated time to keep
FakeJob alive for 1 sec (mocked).

Disallow-Recycled-Builds: test-failures
BUG=b:199714461
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: I5787c80f4c61751e39fee252cb747f7caa7d7151
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3260647
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Vignesh Shenvi <vshenvi@google.com>
Reviewed-by: Kyle Williams <kdgwill@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>