login: Use chrome_setup.cc to set help-content-id switch
Prior to this commit, --device-help-content-id was populated using
serialized-ash-switches in cros_config.  This approach is obviously
much cleaner, but has a disadvantage that serialized-ash-switches is
not available for pre-unibuild (e.g., hana).

This makes the switch set in chrome_setup.cc instead, which is
compatible with both unibuild and pre-unibuild.

BUG=b:203539542
TEST=unit tests pass

Change-Id: Iaa0705f459b4a574f8f38e5cbe3f7ff53ed524ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3257913
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>