Dynamically link camera libs
This CL changes the way camera libraries are linked.

After this CL, camera libraries will be mounted to
/usr/share/cros-camera/libfs on runtime before any usage of the
libraries.

For usages which link the camera libraries .so files while launching the
binary, we specify the "-rpath" linker flag so that linker will also
search for the path where camera libraries are mounted.

For usages which use dlopen() to load camera libraries at runtime, we
only need to change the path it used to the path where camera libraries
are mounted.

BUG=b:193747946
TEST=tast run [DUT] camera.CCAUIDocumentScanning

Cq-Depend: chromium:3259121, chromium:3322512
Change-Id: Idd6d81f9871d441b3053f6d6a4d83439aff291aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3258967
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Tested-by: Wei Lee <wtlee@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>