power: Fix SuspendConfiguratorFuzzer timeout
SuspendConfiguratorFuzzer runs into timeouts when running in a
Clusterfuzz environment. This is because the TaskEnvironment has checks
for the same. We don't really need an entire TaskEnvironment for the
purposes of fuzzing this class (just a TaskExecutor), so instantiate
the same here, so that the SuspendConfigurator tasks can be processed,
and we don't have to deal with the timeout.

BUG=b:204530045
TEST=Run fuzzer for 30 minutes using cros_fuzz shell.

Change-Id: If327143f73016eda96b17f6d045c55be3fb8e238
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3260648
Tested-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>