disk_updater: Add support for BH799
BH799 is not fully nvme compliant as only slot 3 is available for NVMe
firmware update.
Usually, we try to use a firmware slot that is not already in use for
the upgrade, but we cannot with BH799.

Add a quirk to use the same slot.

BUG=b:202985579
TEST=Run 'FEATURES=test emerge-eve-arc-r disk_updater'
[Note to self, using emerge-guybrush triggers SIGILL as expected,
 as AMD code can not run on local workstation.]

Change-Id: I640dbade26fc949d8398ad541fecca15df5d3899
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3260643
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>