shill: vpn: Call StopCharon() in dtor of IPsecConnection
Currently we only call OnDisconnect() in the dtor of IPsecConnection if
the connection is still connected or disconnecting. OnDisconnect() will
only call StopCharon() only if the L2TPConnection does not exist or has
already stopped, and the call of L2TPConnection::Disconnect() does not
have any real effect here since that is executed by PostTask(), and the
L2TPConnection object will be destructed before that. Thus we only need
to do the clean up only for this class in the dtor, so a StopCharon()
should be called instead to make sure that we send SIGTERM to the charon
process correctly.

BUG=None
TEST=tast run $DUT network.VPNConnect.l2tp_ipsec*

Change-Id: I203d35073f567ce4f0674cadd63a838ad4c917b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3261319
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>