lorgnette: Fix missing lock variable
A ScopedClosureRunner created in GetNextImage was taking a lock through
base::AutoLock, but didn't assign the result to a named variable.  This
caused the lock to be released at the end of the expression instead of
protecting the rest of the function.  Give the AutoLock a name so that
it works as expected.

FIXED=b:205131423
BUG=b:205131423
TEST=scan some pages

Change-Id: I35472fa7acb5cf7038ac565402224e9ea40c1e58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3260658
Reviewed-by: Kyra Moed <kmoed@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>