featured: Add a library to check feature state.
This is a reland of 5986530624dcec0e8bb604b78ba92185d7a1a2c1

This fixes the bug in the original CL, which was caused by a test that
verified that we crashed when using distinct feature structs with the
same name. However, that check was a DCHECK, so we should only expect to
crash if DCHECKs are on.

Original change's description:
> featured: Add a library to check feature state.
>
> This library provides a nicer interface around the dbus call
> org.chromium.ChromeFeaturesServiceInterface.state.IsFeatureEnabled,
> providing something closer to what Chrome's base::FeatureList makes
> available.
>
> BUG=b:203674129
> TEST=unit tests
>
> Cq-Depend: chromium:3261097
> Change-Id: I06a58a496b343b3a596fe34f95d98711903d27b4
> Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3260651
> Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
> Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
> Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>

BUG=b:203674129
TEST=unit tests

Cq-Depend: chromium:3283271
Change-Id: If6b8498304f25578dee7ba5f80f4bca28e472f1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3283272
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>