debugd: add a GetPerfOutputV2 function
The V2 version adds a parameter to temporarily disable all CPUs
from entering idle states before collecting perf data. This is
necessary to collect ETM data on ARM devices due to a hardware
bug. The primary functionality is the same as the GetPerfOutputFd
function, which is used in CWP's perf events collector in Chromium.

The added internal functions, one setting the cpuidle disable
value to 1 across all CPUs and states and capturing the
original value, which should almost always be 0. The other
restoring them after finishing the perf command. The state is
stored as a member of the PerfTool instance as it might also
be restored through the StopPerf function to stop a recording
session earlier.

BUG=b:201726173
TEST=FEATURES=test emerge-trogdor debugd && tast run $DUT debugd.PerfTool.*

Change-Id: Ib7863dc77ab15f03b5e046a966b72cae7249d569
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3262985
Reviewed-by: Chinglin Yu <chinglinyu@chromium.org>
Reviewed-by: Gabriel Marin <gmx@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Tuo Shan <shantuo@google.com>
Commit-Queue: Tuo Shan <shantuo@google.com>