Reland "mount-encrypted: Use logical volumes for encstateful"
This is a reland of 53dc5578907c164294ccdd3e42bc6850789912be

The change was originally reverted due to b/206260036: the remove
rootfs verification step writes to the root partition behind the back
of the dm-verity device. Any subsequent reads to the dm-verity device
that landed back at disk caused a kernel panic. LVM commands added in
this CL scanned the dm-verity device and resulted in a panic.

The LVM scanning on the dm-verity device has been disabled with
CL:3292286 which prevents the kernel panic seen above in b/206260036.
Additionally, a longer-term fix for the rootfs verification removal
is tracked in b:207167831, which should make it difficult to trigger
this situation again.

Original change's description:
> mount-encrypted: Use logical volumes for encstateful
>
> If the stateful partition uses LVM, attempt to set up the encrypted
> stateful partition using a logical volume. We use a loopback sparse
> file if the device already has one or if the stateful partition is
> not set up with LVM.
>
> BUG=b:172344853
> TEST=boot with LVM mount-encrypted
>
> Disallow-Recycled-Builds: test-failures
> Change-Id: I5f73106642a5900144ad25174f6ac6d3b452cea5
> Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3263227
> Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
> Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
> Reviewed-by: Gwendal Grignou <gwendal@chromium.org>

BUG=b:172344853, b:207167831
TEST=boot with LVM mount-encrypted.
TEST=cros deploy image, make_dev_ssd succeeds.

Change-Id: Iebd91b6365deb7a2e037fe253459ca6795045d60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3293863
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>