cryptohome: Ensure every AuthBlock return an error when it failed
This change would be helpful to increase the consistency between
asynchronous and synchronous AuthBlocks.

In the asynchronous interface, the error code is more important than
the AuthBlockState. The caller would use the error code to do some
decisions before accessing the AuthBlockState.

In the previous style, it's very easy to forget to modify the error
code when adding new code.
Returning the error explicitly should resolve this problem.

Note: It is obvious that the developer would not forget to return the
AuthBlockState. (At least most of the success-unittest should catch
this case. But it is hard to write unittest for every error case.)

BUG=b:203995025
TEST=unittest of cryptohome
TEST=tast run $DUT hwsec.*

Change-Id: I30d265a2d8f926e897e41473a7eefaa09ee3885c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3263689
Commit-Queue: Yi Chou <yich@google.com>
Tested-by: Yi Chou <yich@google.com>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>