secanomalyd: Send anomalous system reports.
Actual reporting will start when we add the --generate_reports flag to
the secanomalyd command-line, but all the code is done now.

This required a small change to the 'crash_reporter'-calling code to
ensure that the file descriptor used to write the report data was
closed before calling Wait() on the process.

BUG=b:200991907
TEST=Run with --dev, see crash report generated in /var/spool/crash.
TEST=Run crash_sender, get a report id.

Change-Id: I5c8c574e7900849db1dc63f9d26bc40870dee7c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3265204
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>