cryptohome: Remove is pristine from MountHelper
The detection of first mount is to see whether the <hash>/mount/user
directory contains anything that is not created by cryptohome. If it
does - it is not the first mount. Otherwise we assume the previous mount
either never happened or failed. This testing is re-enterable since
nothing is supposed to access cryptohome before it exposes the vault to
designated mount points, and that is the last operation in the sequence.

BUG=b:172344610
TEST=FEATURES="test" USE="test" emerge-hatch cryptohome

Change-Id: I31f0c50a3f1fa6e7e771c83c371988cafab481e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3267442
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Greg Kerr <kerrnel@chromium.org>