patchpanel: Validate ip6->ip6_plen field.
Log errors if there is a mismatch between ip6->ip6_plen and the actual
length of the ip6 packet it says we received.

We don't do an explicit negative ip6_plen check because ntohs returns an
unsigned int{16,32}. current code doesn't ever assign/cast the ip6_plen
to a signed value, so we should never see a negative value. Regardless,
the mismatch check should catch it in case we do accidentally refer to a
signed value.

BUG=b:205304498
TEST=FEATURES="test" emerge-${BOARD} patchpanel

Change-Id: Ia506ea180a4d73330ad9535e0fe307d0d1cc1e92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3272077
Tested-by: Cassie Wang <cassiewang@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Cassie Wang <cassiewang@google.com>