diagnostics: pattern match on wilco device check
Instead of checking for an exact match against a list of wilco devices,
check if the board name string prefix matches any of the listed wilco
devices.

Example: board "sarien-kernelnext" is a wilco device, and "sarien" is a
wilco device, so IsWilcoDevice should return true in this case.

BUG=b:203452507,b:204960820,b:204961289
TEST=cros-health-tool diag –action=run_routine –routine=nvme_wear_level

Change-Id: I4c2e9219acf8c87fcfa90d489eb6c1c0447b1306
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3273194
Tested-by: T.J. Alumbaugh <talumbau@google.com>
Reviewed-by: Kerker Yang <kerker@chromium.org>
Commit-Queue: T.J. Alumbaugh <talumbau@google.com>