shill: net: remove unused IPAddress::GetMinPrefixLength
This patch removes some unused code in IPAddress that was trying to
produce a minimum prefix length based on non-stadard heuristic rules.

The notion of a minimum prefix length other than 0 is not really defined
or meaningful operationally since the prefix length of a network is
always bounded to the wider prefix to which that network is related.

BUG=b:203712717
BUG=b:206055927
TEST=unit tests. Repro case of b/206055927 does not trigger a crash
anymore.

Change-Id: I091d9e684f93186d59895fc8833ed3b78fb08c0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3277705
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>